Found a limitation dealing with duplicat callbacks
Found out that the menu item was how the endpoint was determined. Used that to find out what the current endpoint is. With that, I can do a check immediatly and leave if possible.

Also, this limits the search of the callback function to the one endpoint and not all of the enabled ones. This makes the logic much faster