Checking for blank account...
Doing this at the beginning of the postprocess so that I don't spend any cycles in the api if google analytics wasn't configured for this install.
This usually only happens if someone hasn't set up the account yet in settings or if they override the settings in the settings.php file. This should speed things up a little.